FIx modelstate transfer
